- ottimizzato delete con jquery(no refresh)

+++ app.py
@@ -1,7 +1,7 @@
-from flask import Flask, render_template, request, redirect, url_for
+from flask import Flask, render_template, request, redirect, url_for,jsonify
 import sqlite3
 
-app = Flask(__name__, template_folder='templates')
+app = Flask(__name__, template_folder='templates', static_folder='static')
 
 @app.route("/")
 def form():
@@ -24,15 +24,32 @@ def submit():
     return render_template("form.html",data=data)
 
 
-#Aggiunto il DELETE button
-@app.route("/card/<int:card_id>/delete", methods=["GET"])
+#Aggiunto il DELETE button (vecchio codice con il Refresh)
+# @app.route("/card/<int:card_id>/delete", methods=["GET"])
+# def delete_card(card_id):
+#     conn = sqlite3.connect("database/database.db")
+#     c = conn.cursor()
+#     c.execute("DELETE FROM users WHERE id=?", (card_id,))
+#     conn.commit()
+#     #query to retrieve the data from the table
+#     c.execute("SELECT * FROM users")
+#     data = c.fetchall()
+#     conn.close()
+#     return render_template("form.html",data=data)
+
+
+#CODICE AGGIORNATO PER IL DELETE USANDO JQUERY E JS 
+@app.route('/card/<int:card_id>', methods=['DELETE'])
 def delete_card(card_id):
     conn = sqlite3.connect("database/database.db")
     c = conn.cursor()
-    c.execute("DELETE FROM users WHERE id=?", (card_id,))
-    conn.commit()
-    #query to retrieve the data from the table
-    c.execute("SELECT * FROM users")
-    data = c.fetchall()
-    conn.close()
-    return render_template("form.html",data=data)
+    c.execute("SELECT * FROM users WHERE id=?", (card_id,))
+    result = c.fetchone()
+    if result:
+        c.execute("DELETE FROM users WHERE id=?", (card_id,))
+        conn.commit()
+        conn.close()
+        return jsonify({"message": "Success"}),200
+    else:
+        conn.close()
+        return jsonify({"message": "No item found to delete"}),404
